+++ CHANGELOG.md
@@ -29,6 +29,8 @@ The rules for this file:
 
 <!-- New added features -->
 
+- Added support to duplicate widgets (PR #13)
+
 ### Fixed
 
 <!-- Bug fixes -->